Creating worksheets module, moving code from dataall/api
