+++ backend/dataall/api/Objects/__init__.py
@@ -29,7 +29,6 @@ from . import (
     RedshiftCluster,
     Glossary,
     AthenaQueryResult,
-    Worksheet,
     Feed,
     Notification,
     Vpc,
