+++ backend/dataall/modules/worksheets/api/__init__.py
@@ -1,9 +1,9 @@
-from . import (
+from dataall.modules.worksheets.api import (
     input_types,
     mutations,
     queries,
     resolvers,
     schema,
 )
-# Test change
+
 __all__ = ['resolvers', 'schema', 'input_types', 'queries', 'mutations']
