+++ backend/dataall/modules/worksheets/api/resolvers.py
@@ -1,11 +1,9 @@
-from sqlalchemy import and_
-
-from .... import db
-from ..AthenaQueryResult import helpers as athena_helpers
-from ....api.constants import WorksheetRole
-from ....api.context import Context
-from ....db import paginate, exceptions, permissions, models
-from ....db.api import ResourcePolicy
+from dataall import db
+from dataall.api.Objects.AthenaQueryResult import helpers as athena_helpers
+from dataall.modules.worksheets.api.schema import WorksheetRole
+from dataall.api.context import Context
+from dataall.db import paginate, exceptions, permissions, models
+from dataall.db.api import ResourcePolicy
 
 
 def create_worksheet(context: Context, source, input: dict = None):
